# Don't try to parse non XML value in XPath matcher