--- a/src/main/java/com/github/tomakehurst/wiremock/matching/RequestPattern.java
+++ b/src/main/java/com/github/tomakehurst/wiremock/matching/RequestPattern.java
@@ -78,10 +78,14 @@ public class RequestPattern implements NamedValueMatcher<Request> {
         this.matcher = new RequestMatcher() {
             @Override
             public MatchResult match(Request request) {
-                List<WeightedMatchResult> matchResults = new ArrayList<>(asList(
-                        weight(RequestPattern.this.url.match(request.getUrl()), 10.0),
-                        weight(RequestPattern.this.method.match(request.getMethod()), 3.0),
+                if (!RequestPattern.this.method.match(request.getMethod()).isExactMatch()) {
+                    return MatchResult.noMatch();
+                }
+                if (!RequestPattern.this.url.match(request.getUrl()).isExactMatch()) {
+                    return MatchResult.noMatch();
+                }
 
+                List<WeightedMatchResult> matchResults = new ArrayList<>(asList(
                         weight(allHeadersMatchResult(request)),
                         weight(allQueryParamsMatch(request)),
                         weight(allCookiesMatch(request)),