We try to use Wiremock for performance tests, but there is a bottleneck.
We have a mix of XML and Json stubs, and Wiremock try to parse json request in XML (see issue #776 and #652).
Or parsing data in XML is a heavy operation (CPU intensive, generate exception, ...), so don't try to parse non XML value in XPath matcher

**With this simple test case:**
```java
@Test
public void perfXmlTest() {
	perfRunXPath("matchingXPath on XML datas", "<perf-test>perf-data</perf-test>");
	perfRunXPath("matchingXPath on JSON datas", "{\"perf-test\": \"perf-data\"}");
}
private void perfRunXPath(String message, String value) {
	long ts = System.currentTimeMillis();
	for (int i = 0; i < 100000; i++) {
		WireMock.matchingXPath("//perf-test").match(value);
	}
	System.out.println(message + ":\t" + (System.currentTimeMillis() - ts) + "ms");
}
```
**Without my merge request:**
matchingXPath on XML datas:	19137ms
matchingXPath on JSON datas:	10376ms

**With my merge request:**
matchingXPath on XML datas:	19278ms
matchingXPath on JSON datas:	197ms