@tomakehurst 2019-01-04
Again, thanks for this. I'm definitely keen to improve performance in this area, and clearly given your test results you've found a potentially significant optimisation.

I've added one comment against the code if you could take a quick look.