@philippe-granet 2019-01-04
> I've added one comment against the code if you could take a quick look.

Sorry, which comment ?