@tomakehurst 2019-01-04
Line 83 of `RequestPattern` (it's in this thread, for me at least).